Shelving InputController for now, reorganized the Unit object to have Visuals as a child.
