+++ Unity/Assets/Scenes/Testbeds/MovementTest.unity
@@ -242,10 +242,8 @@ GameObject:
   - component: {fileID: 207405727}
   - component: {fileID: 207405726}
   - component: {fileID: 207405725}
-  - component: {fileID: 207405729}
-  - component: {fileID: 207405728}
   m_Layer: 0
-  m_Name: Unit
+  m_Name: Visuals
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -299,38 +297,57 @@ Transform:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 207405724}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 2
+  m_Father: {fileID: 1014946798}
+  m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &207405728
+--- !u!1 &1014946796
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1014946798}
+  - component: {fileID: 1014946797}
+  m_Layer: 0
+  m_Name: Unit
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!114 &1014946797
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 207405724}
+  m_GameObject: {fileID: 1014946796}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 12c8256ad31b40b4bb31b2e7cb65e273, type: 3}
+  m_Script: {fileID: 11500000, guid: 14e0557d09499664eb74a0d85d9f05f7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   moveDelta: 1
   transitionFrames: 15
   unitMovementDistance: 3
---- !u!114 &207405729
-MonoBehaviour:
+--- !u!4 &1014946798
+Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 207405724}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 1327945023, guid: f70555f144d8491a825f0804e09c671c, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
+  m_GameObject: {fileID: 1014946796}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children:
+  - {fileID: 207405727}
+  m_Father: {fileID: 0}
+  m_RootOrder: 2
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1726639884
 GameObject:
   m_ObjectHideFlags: 0
@@ -437,6 +454,3 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 12c8256ad31b40b4bb31b2e7cb65e273, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  moveDelta: 1
-  transitionFrames: 60
-  unitMovementDistance: 3
